requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
datatypes==0.2.4
decorators==2.0.7
Flask==2.0.1
Flask-JWT-Extended==4.2.3
Flask-Limiter==1.4
gcloud==0.17.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonify==0.5
jws==0.1.3
limits==1.5.1
MarkupSafe==2.0.1
oauth2client==3.0.0
pkg_resources==0.0.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
PyJWT==2.1.0
pyparsing==2.4.7
Pyrebase==3.0.27
python-dotenv==0.19.0
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
